% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_fonts.R
\name{load_default_font}
\alias{load_default_font}
\title{Load Fonts for the 'eastyle' Package}
\usage{
load_default_font()
}
\description{
This function loads fonts that will be used by the \pkg{eastyle} package. Based on the same function from \pkg{showtextdb} by Yixuan Qiu https://statr.me/
}
\examples{
\dontrun{
sysfonts::font_families()

load_showtext_fonts()
sysfonts::font_families()
}
}
\author{
Yixuan Qiu
}
