% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors_pdw.R
\docType{data}
\name{pdw_colors}
\alias{pdw_colors}
\title{Function to extract pdw colors as hex codes}
\format{An object of class \code{character} of length 5.}
\usage{
pdw_colors
}
\arguments{
\item{...}{Character names of pdw_colors}
}
\description{
This function creates the output vector for the colors into hex codes
}
\keyword{datasets}
