% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_color_pdw.R
\name{scale_color_pdw}
\alias{scale_color_pdw}
\title{Color scale constructor for pdw colors}
\usage{
scale_color_pdw(palette = "main", discrete = TRUE, reverse = FALSE, ...)
}
\arguments{
\item{palette}{Character name of palette in pdw_palettes}

\item{discrete}{Boolean indicating whether color aesthetic is discrete or not}

\item{reverse}{Boolean indicating whether the palette should be reversed}

\item{...}{Additional arguments passed to discrete_scale() or
scale_color_gradientn(), used respectively when discrete is TRUE or FALSE}
}
\description{
This function creates the colors from the palette
}
