% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_ggplot2.R
\name{theme_pdw}
\alias{theme_pdw}
\title{Cone Theme Which is Set to the Default}
\usage{
theme_pdw(
  base_size = 14,
  plot_title_size = 18,
  plot_subtitle_size = 16,
  legend.position = "top",
  background_color = "#F9F9F9",
  ...
)
}
\arguments{
\item{base_size}{the base size of the ggplot with default of \code{14}}

\item{plot_title_size}{the size of the title plot with default of \code{18}}

\item{plot_subtitle_size}{the size of the title plot with default of \code{16}}

\item{legend.position}{default of \code{"top"}}

\item{background_color}{the background color of the blot with default of \code{#F9F9F9}}

\item{...}{any additional parameters to pass to \code{ggplot2::theme}}
}
\description{
This function creates the default theme for Cone Graphs
}
\seealso{
Other EA ggplot themes: 
\code{\link{theme_pdw_facet}()}
}
\concept{EA ggplot themes}
