% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_ggplot2.R
\name{theme_pdw_facet}
\alias{theme_pdw_facet}
\title{A Second Cone Theme for Facet Charts}
\usage{
theme_pdw_facet(
  base_size = 10,
  plot_title_size = 18,
  plot_subtitle_size = 16,
  strip_text_size = 12,
  legend.position = "top",
  background_color = "#F9F9F9",
  ...
)
}
\arguments{
\item{base_size}{the base size of the ggplot with default of \code{10}}

\item{plot_title_size}{the size of the title plot with default of \code{18}}

\item{plot_subtitle_size}{the size of the title plot with default of \code{16}}

\item{strip_text_size}{the size of the facet strip text with default of \code{12}}

\item{legend.position}{default of \code{"top"}}

\item{background_color}{the background color of the blot with default of \code{#F9F9F9}}

\item{...}{any additional parameters to pass to \code{ggplot2::theme}}
}
\description{
This theme is designed for use in Facet Charts
called by \code{ggplot2::facet_grid} or \code{ggplot2::facet_wrap}. It adds
color to the strip background and a black outline around the individual
small multiple charts.
}
\seealso{
Other EA ggplot themes: 
\code{\link{theme_pdw}()}
}
\concept{EA ggplot themes}
