% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_pdw_defaults.R
\name{undo_pdw_defaults}
\alias{undo_pdw_defaults}
\title{Resets all the ggplot2 theme to its default settings}
\usage{
undo_pdw_defaults()
}
\description{
Resets all the ggplot2 theme to its default settings
}
